chore : refactor auth system to be cleaner
