+++ backend/src/auth/mod.rs
@@ -1,0 +1,7 @@
+pub mod user;
+pub mod share;
+pub mod middlewares;
+
+pub use user::*;
+pub use share::*;
+pub use middlewares::*;
