--- backend/src/users/mod.rs
@@ -1,7 +1,0 @@
-pub mod auth;
-pub mod model;
-pub mod share;
-pub mod xsrf;
-
-pub use auth::*;
-pub use model::*;
